requirements.txt
pandas
numpy
tensorflow
keras
opencv-python
pydantic
Pillow